matrix-json 2.3.0: remove deprecated JsonImporter/JsonExporter
Removes the legacy JsonImporter/JsonExporter classes (deprecated since
v2.1.2) in favor of JsonReader/JsonWriter, migrates every dependent test
and doc, and bumps the module version. See matrix-json/req/v2.3.0-plan.md
for the full plan.
